Total moving and updates
