--- nikitov.dmitriy/T1/insert_figures.hpp
@@ -1,16 +1,0 @@
-#ifndef INSERT_FIGURES_HPP
-#define INSERT_FIGURES_HPP
-
-#include <string>
-#include "shape.hpp"
-
-namespace nikitov
-{
-  Shape* insertRectangle(std::string line);
-
-  Shape* insertDiamond(std::string line);
-
-  Shape* insertRegular(std::string line);
-}
-
-#endif
